[Decomp][Creature] Cohesion split: extract movement/formation to CreatureMovement.cpp
